requirements.txt
discord.py==1.7.*
environs==9.3.*
